+++ pkg/platformutil/platformutil_test.go
